fix: schema

+++ src/main/resources/schema.sql
@@ -9,7 +9,7 @@ DROP TABLE IF EXISTS friendship_statuses;
 
 CREATE TABLE users
 (
-    id        INT PRIMARY KEY,
+    id        LONG PRIMARY KEY,
     email     VARCHAR(50) UNIQUE,
     login     VARCHAR(20) UNIQUE,
     name      VARCHAR(50),
@@ -19,13 +19,13 @@ CREATE INDEX idx_user_email ON users (email);
 
 CREATE TABLE mpa_ratings
 (
-    id   INT PRIMARY KEY,
+    id   LONG PRIMARY KEY,
     name VARCHAR(50)
 );
 
 CREATE TABLE films
 (
-    id                  INT PRIMARY KEY,
+    id                  LONG PRIMARY KEY,
     name                VARCHAR(50)  NOT NULL,
     description         VARCHAR(250) NOT NULL,
     release_date        DATE         NOT NULL,
@@ -37,37 +37,37 @@ CREATE INDEX idx_film_name ON films (name);
 
 CREATE TABLE likes
 (
-    user_id INT NOT NULL,
-    film_id INT NOT NULL,
+    user_id LONG NOT NULL,
+    film_id LONG NOT NULL,
     FOREIGN KEY (user_id) REFERENCES users (id),
     FOREIGN KEY (film_id) REFERENCES films (id)
 );
 
 CREATE TABLE friendship_statuses
 (
-    id   INT PRIMARY KEY,
+    id   LONG PRIMARY KEY,
     name VARCHAR(50)
 );
 
 CREATE TABLE friendly_relations
 (
-    user_id              INT NOT NULL,
-    friend_id            INT NOT NULL,
-    friendship_status_id INT NOT NULL,
+    user_id              LONG NOT NULL,
+    friend_id            LONG NOT NULL,
+    friendship_status_id LONG NOT NULL,
     PRIMARY KEY (user_id, friend_id),
     FOREIGN KEY (friendship_status_id) REFERENCES friendship_statuses (id)
 );
 
 CREATE TABLE genres
 (
-    id   INT PRIMARY KEY,
+    id   LONG PRIMARY KEY,
     name VARCHAR(50)
 );
 
 CREATE TABLE film_genre
 (
-    film_id  INT NOT NULL,
-    genre_id INT NOT NULL,
+    film_id  LONG NOT NULL,
+    genre_id LONG NOT NULL,
     PRIMARY KEY (film_id, genre_id),
     FOREIGN KEY (film_id) REFERENCES films (id),
     FOREIGN KEY (genre_id) REFERENCES genres (id)
